+++ tests/settings.py
@@ -145,7 +145,6 @@ INBOX_CONFIG = {
             'skip_email': ['group_with_skip_push_3']
         }
     ],
-    'APP_PUSH_NOTIFICATION_KEY_GETTER': 'tests.models.get_notification_key',
     'BACKENDS': {
         'APP_PUSH': 'inbox.core.app_push.backends.locmem.AppPushBackend'
     },